Move delegates to custom packages
