+++ mvvm/src/main/java/com/thefuntasty/mvvm/viewmodel/ViewModelFragmentDelegate.kt
@@ -1,4 +1,4 @@
-package com.thefuntasty.mvvm.delegate
+package com.thefuntasty.mvvm.viewmodel
 
 import androidx.fragment.app.Fragment
 import com.thefuntasty.mvvm.BaseView
